ref: refactor sign up page and button
